+++ packages/008/src/components/Basics.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { 
   ActivityIndicator,
   Image,
